Feat: 페이지에서 사이드바로 연결 변경
